+++ spec/testcentricity_web/elements/label_spec.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-describe TestCentricity::Label, required: true do
+RSpec.describe TestCentricity::Label, required: true do
   subject(:css_label) { described_class.new(:test_label, self, 'label#css_label', :page) }
 
   it 'returns class' do