+++ spec/testcentricity_web/elements/video_spec.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-describe TestCentricity::Video, required: true do
+RSpec.describe TestCentricity::Video, required: true do
   subject(:css_video) { described_class.new(:test_video, self, 'video#css_video', :page) }
 
   it 'returns class' do