+++ spec/testcentricity_web/webdriver_connect/mobile_webdriver_spec.rb
@@ -1,12 +1,12 @@
 # frozen_string_literal: true
 
-describe TestCentricity::WebDriverConnect, mobile: true do
+RSpec.describe TestCentricity::WebDriverConnect, mobile: true do
   before(:context) do
     # instantiate local test environment
     @environs ||= EnvironData
     @environs.find_environ('LOCAL', :yaml)
     ENV['SELENIUM'] = ''
-    ENV['WEB_BROWSER'] = 'appium'
+    ENV['DRIVER'] = 'appium'
     ENV['DEVICE_TYPE'] = 'tablet'
     # start Appium server
     $server = TestCentricity::AppiumServer.new
@@ -21,15 +21,9 @@ describe TestCentricity::WebDriverConnect, mobile: true do
       ENV['APP_VERSION'] = '15.4'
       ENV['APP_DEVICE'] = 'iPad Pro (12.9-inch) (5th generation)'
       WebDriverConnect.initialize_web_driver
-      expect(Environ.browser).to eq(:appium)
-      expect(Environ.platform).to eq(:mobile)
-      expect(Environ.headless).to eq(false)
-      expect(Environ.session_state).to eq(:running)
-      expect(Environ.driver).to eq(:appium)
-      expect(Environ.device).to eq(:simulator)
+      verify_mobile_browser(browser = :safari)
       expect(Environ.device_name).to eq('iPad Pro (12.9-inch) (5th generation)')
       expect(Environ.device_os).to eq(:ios)
-      expect(Environ.device_type).to eq(:tablet)
       expect(Environ.device_os_version).to eq('15.4')
       expect(Environ.is_ios?).to eq(true)
     end
@@ -50,15 +44,9 @@ describe TestCentricity::WebDriverConnect, mobile: true do
         }
       }
       WebDriverConnect.initialize_web_driver(caps)
-      expect(Environ.browser).to eq(:appium)
-      expect(Environ.platform).to eq(:mobile)
-      expect(Environ.headless).to eq(false)
-      expect(Environ.session_state).to eq(:running)
-      expect(Environ.driver).to eq(:appium)
-      expect(Environ.device).to eq(:simulator)
+      verify_mobile_browser(browser = :safari)
       expect(Environ.device_name).to eq('iPad Pro (12.9-inch) (5th generation)')
       expect(Environ.device_os).to eq(:ios)
-      expect(Environ.device_type).to eq(:tablet)
       expect(Environ.device_os_version).to eq('15.4')
       expect(Environ.is_ios?).to eq(true)
     end
@@ -69,15 +57,9 @@ describe TestCentricity::WebDriverConnect, mobile: true do
       ENV['APP_VERSION'] = '12.0'
       ENV['APP_DEVICE'] = 'Pixel_C_API_31'
       WebDriverConnect.initialize_web_driver
-      expect(Environ.browser).to eq(:appium)
-      expect(Environ.platform).to eq(:mobile)
-      expect(Environ.headless).to eq(false)
-      expect(Environ.session_state).to eq(:running)
-      expect(Environ.driver).to eq(:appium)
-      expect(Environ.device).to eq(:simulator)
+      verify_mobile_browser(browser = :chrome)
       expect(Environ.device_name).to eq('Pixel_C_API_31')
       expect(Environ.device_os).to eq(:android)
-      expect(Environ.device_type).to eq(:tablet)
       expect(Environ.device_os_version).to eq('12.0')
       expect(Environ.is_android?).to eq(true)
     end
@@ -98,18 +80,23 @@ describe TestCentricity::WebDriverConnect, mobile: true do
         }
       }
       WebDriverConnect.initialize_web_driver(caps)
-      expect(Environ.browser).to eq(:appium)
-      expect(Environ.platform).to eq(:mobile)
-      expect(Environ.headless).to eq(false)
-      expect(Environ.session_state).to eq(:running)
-      expect(Environ.driver).to eq(:appium)
-      expect(Environ.device).to eq(:simulator)
+      verify_mobile_browser(browser = :chrome)
       expect(Environ.device_name).to eq('Pixel_C_API_31')
       expect(Environ.device_os).to eq(:android)
-      expect(Environ.device_type).to eq(:tablet)
       expect(Environ.device_os_version).to eq('12.0')
       expect(Environ.is_android?).to eq(true)
     end
+  end
+
+  def verify_mobile_browser(browser)
+    expect(Environ.browser).to eq(browser)
+    expect(Environ.platform).to eq(:mobile)
+    expect(Environ.headless).to eq(false)
+    expect(Environ.session_state).to eq(:running)
+    expect(Environ.driver).to eq(:appium)
+    expect(Environ.device).to eq(:simulator)
+    expect(Environ.device_type).to eq(:tablet)
+    expect(Environ.is_web?).to eq(false)
   end
 
   after(:each) do